requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
qdrant-client==1.7.1
sentence-transformers==2.2.2
google-genai==0.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
cross-encoder==0.0.1
pytest==7.4.3
httpx==0.25.2